rename: menu-item-icons→menu-icons
